apps/reframe/backend/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Authentication
python-jose[cryptography]>=3.3.0

# HTTP Client for Upstash Redis REST API
httpx>=0.27.0

# Groq SDK
groq>=0.4.0

# Data validation
pydantic>=2.10.4
pydantic-settings>=2.6.1

# Environment variables
python-dotenv>=1.0.0

# Shared backend package (monorepo)
-e ../../packages/shared-backend  # Local editable install

